Major Changes
I've worked on this assignment for quite a few hours today and have made progress but have also diverted from the original plan for better practicality and hopefully a more fun game. The main goal is no longer to avoid colliding with the asteroids, it's now to shoot as many as you can to prevent your score from reaching 0, when it does you lose.
ERROR
~
Bullet class.